Сделает рендер компонентов с временными данными.
