+++ src/view/edit-point-view.js
@@ -1,0 +1,119 @@
+import {createElement} from '../render';
+import {TYPES_POINTS, DESTINATION} from '../const';
+import {getDateEvent} from '../utils';
+
+function createEditPointTemplate(point, offer, destinations) {
+
+  const {type, basePrice, dateFrom, dateTo} = point;
+  const currentDestination = destinations.find((destination) => destination.id === point.destination);
+  const startEvent = getDateEvent(dateFrom, 'YY/MM/DD HH:mm');
+  const endEvent = getDateEvent(dateTo,'YY/MM/DD HH:mm');
+
+  return (
+    `<li class="trip-events__item">
+      <form class="event event--edit" action="#" method="post">
+      <header class="event__header">
+        <div class="event__type-wrapper">
+          <label class="event__type  event__type-btn" for="event-type-toggle-1">
+            <span class="visually-hidden">Choose event type</span>
+            <img class="event__type-icon" width="17" height="17" src="img/icons/${type}.png" alt="Event type icon">
+          </label>
+          <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
+
+          <div class="event__type-list">
+            <fieldset class="event__type-group">
+              <legend class="visually-hidden">Event type</legend>
+              ${TYPES_POINTS.map((pointType) => (`<div class="event__type-item">
+                <input id="event-type-${pointType}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${pointType}" ${pointType === type ? 'checked' : ' '}>
+                <label class="event__type-label  event__type-label--${pointType}" for="event-type-${pointType}-1">${pointType}</label>
+              </div>`)).join(' ')}
+            </fieldset>
+          </div>
+        </div>
+
+        <div class="event__field-group  event__field-group--destination">
+          <label class="event__label  event__type-output" for="event-destination-1">
+            ${type}
+          </label>
+          <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="${currentDestination.name}" list="destination-list-1">
+          <datalist id="destination-list-1">
+          ${DESTINATION.map((destination) =>`
+          <option value="${destination}"></option>
+          `).join(' ')}
+          </datalist>
+        </div>
+
+        <div class="event__field-group  event__field-group--time">
+          <label class="visually-hidden" for="event-start-time-1">From</label>
+          <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${startEvent}">
+          &mdash;
+          <label class="visually-hidden" for="event-end-time-1">To</label>
+          <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${endEvent}">
+        </div>
+
+        <div class="event__field-group  event__field-group--price">
+          <label class="event__label" for="event-price-1">
+            <span class="visually-hidden">Price</span>
+            &euro;
+          </label>
+          <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="${basePrice}">
+        </div>
+
+        <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
+        <button class="event__reset-btn" type="reset">Delete</button>
+        <button class="event__rollup-btn" type="button">
+          <span class="visually-hidden">Open event</span>
+        </button>
+      </header>
+      <section class="event__details">
+        <section class="event__section  event__section--offers">
+          <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+
+          <div class="event__available-offers">
+          ${offer.slice(Math.floor(Math.random() * 7)).map(({offers}) => offers.map(({title, price}) =>`
+            <div class="event__offer-selector">
+              <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" ${price > 160 ? 'checked' : ''}>
+              <label class="event__offer-label" for="event-offer-luggage-1">
+                <span class="event__offer-title">${title}</span>
+                &plus;&euro;&nbsp;
+                <span class="event__offer-price">${price}</span>
+              </label>
+            </div>
+          `)).join('')}
+          </div>
+        </section>
+
+        <section class="event__section  event__section--destination">
+          <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+          <p class="event__destination-description">${destinations.map(({description}) => description).join(' ')}</p>
+        </section>
+      </section>
+      </form>
+      </li>`
+  );
+}
+
+export default class EditPointView {
+
+  constructor(points, offers, destinations) {
+    this.points = points;
+    this.offers = offers;
+    this.destinations = destinations;
+  }
+
+  getTemplate() {
+    return createEditPointTemplate(this.points, this.offers, this.destinations);
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
